feat: 🎸 useMergedState and useTogglable
useMergedState take an initial state (which must be an object) and
return a typle of state + state updater which receive an object and
merge it to the current state. useTogglable take a boolean and return a
tuple of the current state of the toggled state, and a function that
will toggle the current state.